Improve INSTALL doc instructions
1. Update cmdline to cater for yasm dependency
qatlib build from sources depends on yasm since 22.07 release.
(This is temporary, will be removed in a future release)
So adding installation of yasm to command line instructions.

2. Refine steps to help detect missing firmware
Include xz file format.
Add instruction to run dracut on any firmware file update.
Add extra prerequisites step.
Add that missing fw issue would be seen on running systemctl start qat.
Add missing "sudo" on running dracut.

3. Make clear that VFs are not a prerequisite.
VFs will get created during the installation process if they're
not already there from a previous installation. Prerequisites
text is adjusted to reflect that.
Also the PF text is adjusted so it will be easier to add new devices.

4. Differentiate between lib and sample_code steps
Users may want to use the library with their own application, so make
clearer which steps are necessary for installation of the lib for use with
any application and which are specific to running the provided sample
application.

Signed-off-by: Fiona Trahe <fiona.trahe@intel.com>