Update in documentation - in INSTALL file
Update in the description for the default configuration
in the Configuration and Tuning section. A correction
is made in the maximum number of processes.

As for POLICY 0, the maximum number of processes will be always 16
regardless of number of sockets!

i.e. for POLICY = 0, each process gets 1 VF from each PF;
for 1S we have 16 VFs * 4 PFs = 64 VFs; one process can take 4 VF
so a maximum of 16 (64 / 4) number of processes are possible;
for 2S we have 16 VFs * 8 PFs = 128 VF; so a maximum of
16 (128 / 8) number of processes are possible.

Correction is made based on the details above.

Signed-off-by: Firos Sadarul <firos.sadarul@intel.com>