[DOC] QATlib README update to add 402xx/SRF support
README doc update as the 402xx device has been validated on the
24.02 pkg post-release.

Add/update links to point to the QAT public wiki docs.

Also explicitly list the 401xx in the supported devices list. This was
already supported and documented in 22.07 (CPM2.0b support), but not
explicitly listed, as included under the 4xxx driver. Explicitly listing
now for clarity.

Signed-off-by: Fiona Trahe <fiona.trahe@intel.com>
Signed-off-by: Meadhbh Fitzpatrick <meadhbh.fitzpatrick@intel.com>