Re-enable initialization in the absence of PFVF comms
This change is to allow QATlib to successfully initialize and
allocate resources to a process when running on an older kernel
like LTS 5.15 where PF/VF comms is not fully functional. In such, a
situation, we would like a client process to be initialized
on the basis of a default configuration.

This is restoring the functionality which was included in previous
releases but accidentally removed in the previous commit.

Signed-off-by: Ahsan Atta <ahsan.atta@intel.com>
Signed-off-by: Fiona Trahe <fiona.trahe@intel.com>
Signed-off-by: Giovanni Cabiddu <giovanni.cabiddu@intel.com>